libft updated

+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: grota <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/20 13:02:13 by grota             #+#    #+#             */
-/*   Updated: 2017/11/20 13:02:14 by grota            ###   ########.fr       */
+/*   Updated: 2017/11/22 12:06:13 by grota            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (i < n && src[i])
